requirements.txt
openai==1.55.3
langchain==0.3.7
langchain-openai==0.2.9
python-dotenv==1.0.1
streamlit==1.41.1